+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - Cache images with dependencies installed, a lot of data saved on pulled dependencies [#38](https://github.com/wojciechkepka/pkger/pull/38)
 - Dependencies now use the TOML syntax instead of a custom one [#39](https://github.com/wojciechkepka/pkger/pull/39)
 - Commands in configure, build and install scripts now use TOML syntax [#40](https://github.com/wojciechkepka/pkger/pull/40)
+- Add `--trace` option that sets log level of pkger to trace and make `--debug` actually set debug